pyproject.toml
[project]
name = "leaderboard-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jq>=1.8.0",
    "pydantic-settings>=2.6.1",
    "pydantic>=2.9.2",
    "requests>=2.32.3",
    "twitter-api-client>=0.10.22",
]
